Modify _load_model function structure
